Update to v6.1.2.0-82
Disable stripping symbols from binaries and libraries. This must be
disabled, otherwise the library files won't pass the verification check
that GlobalProtect does.

Change gpa.service to be a user service as that should (at least with
this version) be run as the user.

Include the globalprotect.desktop file.

Remove the GlobalProtect_tar source file, as I don't have that available
and it's not necessary.